+++ CHANGELOG.md
@@ -33,7 +33,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * trove4j 3.1.0
   * javax.validation-api 2.0.0.Final
 * @NotEmpty now evaluates Map/Collection/Array based on their size/length properties
-
+* renamed PreCheck/PostCheck#getExpression to PreCheck/PostCheck#getExpr
+* renamed PreCheck/PostCheck#getLanguage to PreCheck/PostCheck#getLang
+ 
 
 ### Removed
 * net.sf.oval.integration.spring.BeanInjectingCheckInitializationListener - use net.sf.oval.integration.spring.SpringCheckInitializationListener instead